+++ makemake/core/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from makemake.core.models import UserLog
 
-# Register your models here.
+@admin.register(UserLog)
+class UserLogAdmin(admin.ModelAdmin):
+    list_display = ("user", "login_time", "last_login_ip")
+    search_fields = ("user__username", "last_login_ip")