+++ makemake/core/models.py
@@ -1,3 +1,10 @@
 from django.db import models
+from django.contrib.auth.models import User
 
-# Create your models here.
+class UserLog(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="userlog")
+    login_time = models.DateTimeField(auto_now_add=True)
+    last_login_ip = models.GenericIPAddressField(null=True, blank=True)
+
+    def __str__(self):
+        return f"Login de {self.user.username} em {self.login_time} com IP {self.last_login_ip}"